Refactored according to issue #36
